Added RPY to delta-position, and moved orientation math
