+++ aquadrone_sim_worlds/worlds/empty_underwater_timed.world
@@ -33,9 +33,9 @@
       <cast_shadows>false</cast_shadows>
     </light>
     <physics default="true" name="default_physics" type="ode">
-      <max_step_size>0.02</max_step_size>
+      <max_step_size>0.002</max_step_size>
       <real_time_factor>1.0</real_time_factor>
-      <real_time_update_rate>50.0</real_time_update_rate>
+      <real_time_update_rate>500</real_time_update_rate>
       <ode>
         <solver>
           <type>quick</type>
